Added conirmation to delete_env.sh

+++ delete_env.sh
@@ -5,5 +5,16 @@ if [ -n "$1" ]; then
     user_name="$1"
 fi
 
-userdel ${user_name}
-rm -rf /home/${user_name}/
+echo "WARNING! THIS SCRIPT WILL REMOVE USER ${user_name} AND IT'S HOMEDIR (/home/${user_name}/) WITH TEMPEST REPORTS, PREPARED VENV ETC."
+read -p "Are you sure you want to continue? <yes/no> " prompt
+
+if [[ $prompt =~ [yY](es)* ]]
+then
+    userdel ${user_name}
+    rm -rf /home/${user_name}/
+else
+  exit 0
+fi
+
+
+
